% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimpleRegressionData.R
\name{SimpleRegressionData}
\alias{SimpleRegressionData}
\title{Generates sample data for a simple regression, y = alpha + beta x + epsilon with control over var(x) and var(epsilon)}
\usage{
SimpleRegressionData(
  alpha = 1,
  beta = 5,
  rangex = c(0, 10),
  sigma = 1,
  N = 100
)
}
\arguments{
\item{alpha}{intercept of regression.}

\item{beta}{slope coefficient.}

\item{rangex}{range of the uniform normal that will be used to generate x variables in the form c(max,min).}

\item{sigma}{standard deviation of the error term}

\item{N}{Number of observations}
}
\value{
Data.frame with Y, X and er.
}
\description{
Generates sample data for a simple regression, y = alpha + beta x + epsilon with control over var(x) and var(epsilon)
}
\examples{
SimpleRegressionData(N =20)
SimpleRegressionData(alpha = 2, beta = -2, rangex= c(-10, 20), sigma = 3, N =100)
}
